+++ CHANGELOG.md
@@ -17,6 +17,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
+- Fixed annotation details mask editing to keep focus stable without annoying recentering after every edit and always select the shown annotation.
+
 ### Security
 
 ## \[0.4.8\] - 2026-02-11